Added "Home" scene and the "Log out" button.
