+++ src/main/java/org/CarLounge/fis/controller/HomeController.java
@@ -1,0 +1,23 @@
+package org.CarLounge.fis.controller;
+
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.control.PasswordField;
+import javafx.scene.control.TextField;
+import javafx.scene.text.Text;
+import javafx.scene.control.Button;
+import javafx.scene.input.MouseEvent;
+import javafx.stage.Stage;
+
+public class HomeController {
+    @FXML
+    Button logoutfromacc;
+    public void backToLogin(MouseEvent event) throws Exception{
+
+        Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("login.fxml"));
+        Stage window = (Stage)logoutfromacc.getScene().getWindow();
+        window.setScene(new Scene(root));
+    }
+}
