+++ src/main/java/org/CarLounge/fis/services/UserService.java
@@ -18,7 +18,7 @@ public class UserService {
 
     public static void initDatabase() {
         Nitrite database = Nitrite.builder()
-                .filePath(getPathToFile("registration-example.db").toFile())
+                .filePath(getPathToFile("CarLoungeUsers.db").toFile())
                 .openOrCreate("test", "test");
 
         userRepository = database.getRepository(User.class);
