+++ src/main/resources/home.fxml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.CarLounge.fis.controller.HomeController">
+   <children>
+      <Button fx:id="logoutfromacc" layoutX="266.0" layoutY="188.0" mnemonicParsing="false" onMouseClicked="#backToLogin" text="Log out" textAlignment="CENTER" />
+   </children>
+</AnchorPane>
